+++ src/app.py
@@ -22,7 +22,7 @@ def create_app(context: AppContext) -> FastAPI:
         openapi_url="/documentation/json",
         redoc_url=None,
         title="ai-rag-template",
-        version="0.5.1"
+        version="0.5.2"
     )
 
     app.add_middleware(AppContextMiddleware, app_context=context)