+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- At service startup, if the Vector Search collection does not exist, it is automatically created
 - Support file extension `mdx` for embedding generation
 - File uploaded for embedding generation is validated either from the content-type or the file extension