+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.5.2 - 2025-01-29
+
 ### Fixed
 
 - At service startup, if the Vector Search collection does not exist, it is automatically created